Reworked m2m signal handling w/ versioning.
Before, we might sometimes add two versions of the same model, m, because we
would call hist_through.add(m_h) where m_h is the historical instance.  But
we wouldn't check to see that m_h wasn't already in the set with a different
version.

We now ensure that we don't double-add to the set.