Region explore page independent map.
On mobile, we need to hide the map on the region explore page because,
otherwise, it takes up too much vertical space and people don't know how
to either 1) move away from it onto other content 2) won't be able to
interact with it if we ignore touch events on the map surface.

So, we hide the map and provide a map toggle for small screens.