Kill off "Templates/" pages, for now.
Based on watching lots and lots of new users, "Templates" are deeply
confusing, as currently implemented.  In Oakland, the Template pages
are never used for creating new pages, and they've seen a lot of success
with this largely unstructured approach.

We'll reevaluate using some form of "templates" when we introduce the
infobox branch, but until then, it's just too confusing for the marginal
benefit it introduces.