Don't allow Front Page rename/map edit.
Renamign the Front Page doesn't work correctly right now and leads to a
lot of confusion, so it's now disabled.

Editing the map on the Front Page is also confusing, because people
usually just want to edit the Region Settings map.  Maybe we should make
the map edit button a link to that Region Edit area, but that's also
kind of confusing.  So for now this is also disabled.