Work on `fix_tags` utility function.
This is a function that will be used to fix `PageTagSet` objects who
don't, for whatever reason, point at `Tag` objects inside the TagSet
`region`. This can happen during a `move_to_region()`.

In the future, this utility function will be useless as `Tag` will
be a global, not regional, object.