Not sure of why, but the {% with %} tag was causing issues.
Removing the usage of the {% with %} tag here fixes an odd issue where
the contents of the version history shown on the Info page would
randomly display unrelated page history.  It looks like a possible
threading / concurrency issue, or perhaps something wrong with how
caching (maybe at the template level?) works with the with tag here.

I don't have time to investigate this further right now, but this
appears to fix the issue.